# src/spec/management-api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API to be used by Concerto CMS",
        "version": "1.0.0",
        "title": "Concerto Management API"
    },
    "host": "api.concerto-cms-org",
    "basePath": "/mgmt/v1",
    "schemes": [
        "https"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "x-swagger-router-controller": "profile",
                "operationId": "getProfile",
                "summary": "User Profile",
                "description": "The User Profile endpoint returns information about the Uber user that has authorized with the application.",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Profile information for a user",
                        "schema": {
                            "$ref": "#/definitions/Profile"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/sites": {
            "get": {
                "tags": [
                    "Sites"
                ],
                "summary": "Sites available for the user",
                "description": "Retrieve a list of sites that are available to the user given a certain role\n",
                "parameters": [],
                "operationId": "getAvailableSites",
                "x-swagger-router-controller": "sites",
                "responses": {
                    "200": {
                        "description": "An array of sites",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Site"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new site",
                "tags": [
                  "Sites"
                ],
                "description": "Create a new site (where the user will automatically be granted the OWNER role",
                "parameters": [

                ],
                "x-swagger-router-controller": "sites",
                "operationId": "createSite",
                "responses": {
                    "200": {
                        "description": "Site created successfully",
                        "schema": {
                            "$ref": "#/definitions/Site"
                        }
                    }
                }
            }
        },
        "/sites/{siteID}": {
            "parameters": [
                {
                    "name": "siteID",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Sites"
                ],
                "x-swagger-router-controller": "sites",
                "operationId": "getSite",
                "summary": "Site details",
                "responses": {
                    "200": {
                        "description": "Details of the given Site",
                        "schema": {
                            "$ref": "#/definitions/Site"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Sites"
                ],
                "x-swagger-router-controller": "sites",
                "operationId": "updateSite",
                "summary": "Update a site",
                "responses": {
                    "200": {
                        "description": "Details of the given Site",
                        "schema": {
                            "$ref": "#/definitions/Site"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sites"
                ],
                "x-swagger-router-controller": "sites",
                "operationId": "removeSite",
                "summary": "Remove a site",
                "responses": {
                    "204": {
                        "description": "Site has been deleted"
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/sites/{siteID}/blocks": {
            "parameters": [
                {
                    "name": "siteID",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Blocks"
                ],
                "summary": "List of blocks",
                "description": "Retrieve a list of blocks\n",
                "parameters": [],
                "x-swagger-router-controller": "blocks",
                "operationId": "getBlocks",
                "responses": {
                    "200": {
                        "description": "An array of blocks",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Block"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new block",
                "tags": [
                    "Blocks"
                ],
                "description": "Create a new block",
                "parameters": [],
                "x-swagger-router-controller": "blocks",
                "operationId": "createBlock",
                "responses": {
                    "200": {
                        "description": "Block created successfully",
                        "schema": {
                            "$ref": "#/definitions/Block"
                        }
                    }
                }
            }
        },
        "/sites/{siteID}/blocks/{blockID}": {
            "parameters": [
                {
                    "name": "siteID",
                    "in": "path",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "blockID",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Blocks"
                ],
                "x-swagger-router-controller": "blocks",
                "operationId": "getBlock",
                "summary": "Block details",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Details of the given Block",
                        "schema": {
                            "$ref": "#/definitions/Block"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Blocks"
                ],
                "x-swagger-router-controller": "blocks",
                "operationId": "updateBlock",
                "summary": "Update a block",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Block updated",
                        "schema": {
                            "$ref": "#/definitions/Block"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Blocks"
                ],
                "x-swagger-router-controller": "blocks",
                "operationId": "removeBlock",
                "summary": "Remove a site",
                "parameters": [],
                "responses": {
                    "204": {
                        "description": "Block has been deleted"
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/sites/{siteID}/models": {
            "parameters": [
                {
                    "name": "siteID",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Models"
                ],
                "summary": "List of models",
                "description": "Retrieve a list of models\n",
                "parameters": [],
                "operationId": "getAvailableModels",
                "x-swagger-router-controller": "models",
                "responses": {
                    "200": {
                        "description": "An array of models",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Model"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new model",
                "tags": [
                    "Models"
                ],
                "description": "Create a new model",
                "parameters": [],
                "x-swagger-router-controller": "models",
                "operationId": "createModel",
                "responses": {
                    "200": {
                        "description": "Model created successfully",
                        "schema": {
                            "$ref": "#/definitions/Model"
                        }
                    }
                }
            }
        },
        "/sites/{siteID}/models/{modelID}": {
            "parameters": [
                {
                    "name": "siteID",
                    "in": "path",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "modelID",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Models"
                ],
                "x-swagger-router-controller": "models",
                "operationId": "getModel",
                "summary": "Model details",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Details of the given model",
                        "schema": {
                            "$ref": "#/definitions/Model"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Models"
                ],
                "x-swagger-router-controller": "models",
                "operationId": "updateModel",
                "summary": "Update a model",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Model updated",
                        "schema": {
                            "$ref": "#/definitions/Model"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Models"
                ],
                "x-swagger-router-controller": "models",
                "operationId": "removeModel",
                "summary": "Remove a model",
                "parameters": [],
                "responses": {
                    "204": {
                        "description": "Model has been deleted"
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/sites/{siteID}/models/{modelID}/content": {
            "parameters": [
                {
                    "name": "siteID",
                    "in": "path",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "modelID",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Content"
                ],
                "summary": "List of content items",
                "description": "Retrieve a list of content items\n",
                "parameters": [],
                "x-swagger-router-controller": "content",
                "operationId": "getContent",
                "responses": {
                    "200": {
                        "description": "An array of content items",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Content"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new content item",
                "tags": [
                    "Content"
                ],
                "description": "Create a new content item",
                "parameters": [],
                "x-swagger-router-controller": "content",
                "operationId": "createItem",
                "responses": {
                    "200": {
                        "description": "Content item created successfully",
                        "schema": {
                            "$ref": "#/definitions/Content"
                        }
                    }
                }
            }
        },
        "/sites/{siteID}/content": {
            "parameters": [
                {
                    "name": "siteID",
                    "in": "path",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "modelID",
                    "in": "query",
                    "required": false,
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Content"
                ],
                "summary": "List of content items",
                "description": "Retrieve a list of content items\n",
                "parameters": [],
                "x-swagger-router-controller": "content",
                "operationId": "getContent",
                "responses": {
                    "200": {
                        "description": "An array of content items",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Content"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new content item",
                "tags": [
                    "Content"
                ],
                "description": "Create a new content item",
                "parameters": [],
                "x-swagger-router-controller": "content",
                "operationId": "createItem",
                "responses": {
                    "200": {
                        "description": "Content item created successfully",
                        "schema": {
                            "$ref": "#/definitions/Content"
                        }
                    }
                }
            }
        },
        "/sites/{siteID}/content/{contentID}": {
            "parameters": [
                {
                    "name": "siteID",
                    "in": "path",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "modelID",
                    "in": "path",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "contentID",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Content"
                ],
                "x-swagger-router-controller": "content",
                "operationId": "getContentItem",
                "summary": "Block details",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Details of the given item",
                        "schema": {
                            "$ref": "#/definitions/Content"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Content"
                ],
                "x-swagger-router-controller": "content",
                "operationId": "updateContentItem",
                "summary": "Update an item",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Content item updated",
                        "schema": {
                            "$ref": "#/definitions/Content"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Content"
                ],
                "x-swagger-router-controller": "content",
                "operationId": "removeItem",
                "summary": "Remove a content item",
                "parameters": [],
                "responses": {
                    "204": {
                        "description": "Content item has been deleted"
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Site": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Unique identifier representing a specific site."
                },
                "name": {
                    "type": "string",
                    "description": "Name of the site."
                },
                "role": {
                    "type": "string",
                    "description": "Role of the current user"
                },
                "users": {
                    "type": "array",
                    "description": "List of users contributing to the site.",
                    "items": {
                        "$ref": "#/definitions/SiteUser"
                    }
                }
            }
        },
        "SiteUser": {
            "type": "object",
            "properties": {
                "userId": {
                    "type": "string",
                    "description": "Unique identifier of the user"
                },
                "role": {
                    "type": "string",
                    "description": "Role of the given user for the site"
                }
            }
        },
        "Profile": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Unique identifier of the user"
                },
                "username": {
                    "type": "string",
                    "description": "Username"
                },
                "name": {
                    "type": "string",
                    "description": "(Nick)name of the user"
                },
                "picture": {
                    "type": "string",
                    "description": "Picture taken from gravatar or social account"
                },
                "email": {
                    "type": "string",
                    "description": "Email address of the user"
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "Block": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "template": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/BlockField"
                    }
                }
            }
        },
        "BlockField": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "Model": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "fields": {
                    "type": "array"
                }
            }
        },
        "Content": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        }
    }
}